State the intended integer type at the remaining mixed-width call sites
The same deduction failures as before, in the code that had not been reached yet: min, max and
div_rounding_up_safe given one int literal and one size_type, a lambda or ternary whose two
arms disagree once the types diverge, and a brace initializer that becomes narrowing. Nothing
here changes what is computed.

The cooperative-groups reduction in strings::slice is the one worth noting: it was summing
size_type character counts with a plus<int>, which cg::reduce rejects outright rather than
silently narrowing.

+++ cpp/include/cudf/strings/detail/convert/fixed_point_to_string.cuh
@@ -24,7 +24,7 @@ __device__ inline int32_t fixed_point_string_size(__int128_t const& value, int32
   auto const abs_value = numeric::detail::abs(value);
   auto const exp_ten   = numeric::detail::exp10<__int128_t>(-scale);
   auto const fraction  = count_digits(abs_value % exp_ten);
-  auto const num_zeros = cuda::std::max(0, (-scale - fraction));
+  auto const num_zeros = cuda::std::max(0, static_cast<int>(-scale - fraction));
   return static_cast<int32_t>(value < 0) +    // sign if negative
          count_digits(abs_value / exp_ten) +  // integer
          1 +                                  // decimal point
@@ -57,7 +57,8 @@ __device__ inline void fixed_point_to_string(__int128_t const& value, int32_t sc
   if (value < 0) *out_ptr++ = '-';  // add sign
   auto const abs_value = numeric::detail::abs(value);
   auto const exp_ten   = numeric::detail::exp10<__int128_t>(-scale);
-  auto const num_zeros = cuda::std::max(0, (-scale - count_digits(abs_value % exp_ten)));
+  auto const num_zeros =
+    cuda::std::max(0, static_cast<int>(-scale - count_digits(abs_value % exp_ten)));
 
   out_ptr += integer_to_string(abs_value / exp_ten, out_ptr);  // add the integer part
   *out_ptr++ = '.';                                            // add decimal point
